pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "twine"]
build-backend = "setuptools.build_meta"

[project]
name = "dlwheel"
version = "1.0.4"
description = "A lightweight deep learning library."
readme = "README.md"
authors = [{ name = "liiuhaao", email = "liiuhaao66@gmail.com" }]
requires-python = ">=3.6"
dependencies = ["pyyaml", "pathspec"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
